refactor: move InstallDir to github package
